// src/types/CarsTypes.ts
export interface CarTypes {
    brand: string;
    model: string;
    year: number;
    slug:string;
    price: number;
    seat: number;
    doors: number;
    color: string;
    fuelType: string;
    imageUrl: string;
    images: string[];
    description: string;
    availability: boolean;
    bag: number;
    trunk: string;
    km: string;
    airbag: boolean;
    navigation: string;
    transmission: string;
    moteur: string;
    acceleration: string;
    consommationCarburant: string;
    features: string[];
}